Adds hadling of secure hosts in `_createEdgeHostname` and `create`.
This change renames the `secure` parameter to `certEnrollmentId`.
If a certEnrollmentId is given the created hostname will made to use the and the edgehostname will be suffixed with `edgekey.net` instead of `edgesuite.net`.

`create()` is also given a new parameter `certEnrollmentId` to allow library users to use this new feature.